+++ utopia-remix/app/routes-test/internal.projects.destroy.spec.ts
@@ -7,7 +7,7 @@ import {
   newTestRequest,
   truncateTables,
 } from '../test-util'
-import { ApiError } from '../util/api.server'
+import { ApiError } from '../util/errors'
 
 describe('handleDestroyAllProjects', () => {
   afterEach(async () => {
