+++ utopia-remix/app/routes/internal.projects.$id.destroy.tsx
@@ -1,7 +1,7 @@
 import { ActionFunctionArgs } from '@remix-run/node'
 import { Params } from '@remix-run/react'
 import { ensure, handle, requireUser } from '../util/api.server'
-import { Status } from '../util/statusCodes.server'
+import { Status } from '../util/statusCodes'
 import { hardDeleteProject } from '../models/project.server'
 
 export async function action(args: ActionFunctionArgs) {
