+++ utopia-remix/app/routes/internal.projects.$id.rename.tsx
@@ -2,7 +2,7 @@ import { ActionFunctionArgs } from '@remix-run/node'
 import { renameProject } from '../models/project.server'
 import { ensure, handle, requireUser } from '../util/api.server'
 import slugify from 'slugify'
-import { Status } from '../util/statusCodes.server'
+import { Status } from '../util/statusCodes'
 import { Params } from '@remix-run/react'
 
 export const SLUGIFY_OPTIONS = { lower: true, remove: /[^a-z0-9A-Z ]/ }
